more msvc fixes. make dxViewer working.

+++ opensubdiv/osd_util/batch.h
@@ -255,7 +255,7 @@ OsdUtilMeshBatchBase<DRAW_CONTEXT>::createPatchArrays(std::vector<FarPatchTables
 
         // XXX: temporary for uniform
         int maxLevel = meshVector[i]->GetSubdivisionTables()->GetMaxLevel();
-        int numQuads = adaptive ? 0 : meshVector[i]->GetFaceVertices(maxLevel-1).size()/4;
+        int numQuads = adaptive ? 0 : (int)meshVector[i]->GetFaceVertices(maxLevel-1).size()/4;
         _quadCounts.push_back(numQuads);
 
         vertexOffset += numVertices;
